feat: CRUD 로직 변경

+++ src/monthly/monthly.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { CreateMonthlyDto, UpdateMonthlyDto } from './dto/monthly.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Monthly } from '../domain/monthly.entity';
-import { Repository } from 'typeorm';
+import { DeepPartial, Repository } from 'typeorm';
 
 @Injectable()
 export class MonthlyService {
@@ -11,8 +11,11 @@ export class MonthlyService {
     private monthlyRepository: Repository<Monthly>,
   ) {}
 
-  findAll(): Promise<Monthly[]> {
-    return this.monthlyRepository.find();
+  findAll(userEmail: string): Promise<Monthly[]> {
+    return this.monthlyRepository
+      .createQueryBuilder('monthly')
+      .where('monthly.userId = :userEmail', { userEmail })
+      .getMany();
   }
 
   findOne(id: number): Promise<Monthly> {
@@ -20,20 +23,22 @@ export class MonthlyService {
   }
 
   async create(monthly: CreateMonthlyDto): Promise<Monthly[]> {
-    await this.monthlyRepository.save(monthly);
-    return this.monthlyRepository.find();
+    await this.monthlyRepository.save(monthly as DeepPartial<Monthly>);
+    return this.findAll(monthly.userId);
   }
 
   async update(monthly: UpdateMonthlyDto): Promise<Monthly[]> {
     const existed = await this.findOne(monthly.id);
     if (!existed) throw new Error('Monthly not found');
     const nextMonthly = { ...existed, ...monthly };
-    await this.monthlyRepository.save(nextMonthly);
-    return this.monthlyRepository.find();
+    await this.monthlyRepository.save(nextMonthly as DeepPartial<Monthly>);
+    return this.findAll(monthly.userId);
   }
 
   async remove(id: number): Promise<Monthly[]> {
+    const existed = await this.findOne(id);
+    if (!existed) throw new Error('Monthly not found');
     await this.monthlyRepository.delete(id);
-    return this.monthlyRepository.find();
+    return this.findAll(existed.userId);
   }
 }
